move alpha_mask to pixels
