+++ src/lib.rs
@@ -134,7 +134,6 @@ extern crate log;
 #[doc(hidden)]
 pub use freetype as ft;
 
-pub mod alphamask;
 pub mod base;
 pub mod clip;
 pub mod color;
@@ -142,7 +141,7 @@ pub mod line_interp;
 pub mod outline;
 pub mod outline_aa;
 pub mod paths;
-pub mod pixfmt;
+pub mod pixels;
 pub mod ppm;
 pub mod raster;
 pub mod render;
@@ -156,8 +155,6 @@ pub(crate) mod scan;
 
 pub mod gallery;
 
-#[doc(hidden)]
-pub use crate::alphamask::*;
 #[doc(hidden)]
 pub use crate::base::*;
 #[doc(hidden)]
@@ -173,7 +170,7 @@ pub use crate::outline_aa::*;
 #[doc(hidden)]
 pub use crate::paths::*;
 #[doc(hidden)]
-pub use crate::pixfmt::*;
+pub use crate::pixels::*;
 #[doc(hidden)]
 pub use crate::raster::*;
 #[doc(hidden)]
