+++ src/pixels/buffer.rs
