+++ src/pixels/cover.rs
